Change to keeping track of things using map
The array method I was previously using required an iteration through the array for every segment.
So to find the second segment you would go through the entire array, and to find the third segment you would need to go through it again.

The map method doesn’t necessarily fix this but it does make the lookups faster and the code more readable.

Signed-off-by: Nathan Minissale <nathanminissale@gmail.com>